+++ wazo_call_logd/plugins/cdr/schemas.py
@@ -208,6 +208,10 @@ class CDRListRequestSchema(CDRListingBase):
         load_default=None,
         attribute='requested_internal_exten',
     )
+    requested_internal_context = fields.String(
+        validate=Length(min=1, max=255),
+        load_default=None,
+    )
 
     @post_load
     def map_order_field(self, in_data, **kwargs):